+++ src/lib/index.ts
@@ -1,1 +1,3 @@
 export * from "./registerUpdatePrompt.js";
+export * from "./registerServiceWorker.js";
+export * from "./setupServiceWorker.js";
