+++ src/lib/registerServiceWorker.ts
@@ -2,8 +2,8 @@ import { Workbox } from "workbox-window";
 import {
   PermissionCallback,
   askNotificationPermission,
-} from "./notifications/browser.js";
-import { registerUpdatePrompt } from "./lib/registerUpdatePrompt.js";
+} from "../notifications/browser.js";
+import { registerUpdatePrompt } from "./registerUpdatePrompt.js";
 
 export type SetupBrowserOptions = {
   /** url to check for the serviceWorker */
@@ -55,7 +55,7 @@ export function registerServiceWorker(
         wb.addEventListener("waiting", (_event) => {
           console.debug(
             `A new service worker has installed, but it can't activate` +
-              `until all tabs running the current version have fully unloaded.`,
+            `until all tabs running the current version have fully unloaded.`,
           );
         });
       }
